Initial GOAP Project
Current work on the project.

Base Core planning is working.

Includes an example simulation of Chefs creating a plan of actions to fill a plate with the desired ingredients.

2 forms of planning is in place. First a plate spawns and plans the target actions required to fill the recipe (eg. needs 2 buns so the plan is add bun, add bun).
The chefs use these actions as their desired goal to fill the plate.(eg. chef wants to add bun therefore the plan is go to bun station, pick up bun, go to plate, put in plate).