Planner Visualiser Window
Created a planner visualiser for clearer editing.
This is currently targeting the plateRecipeActions, NOT the chefAgentActions.
The window shows how the tree would branch out based on the action collection available.
Each action tab in the planner tree shows what that current world state would look like based on the actions used to get there.