Pruning Search trees. + Eatposition fix
GOAPPlannerObject now prunes future states if it is found to equal an already existing state. The method to determine if it is an equal state is to check if the actions that lead to the current planning state are an equal count. (eg. 2 up buns and an up lettuce is the same as 1 up bun, 1 up lettuce and then another 1 up bun).

I've left some methods in to choose to prune or not. It may be the user's plan to use more information than just the successful branches.

EatingPosition no longer contains the handle drawing. Handles needed to use the unityEditor namespace which prevents builds from compiling. Now there is a CustomEditor for EatingPositions.